Tests: remove duplication of tmpfile fixture
Could also be useful in other unit tests in the future.
